Change some things back to the way they were; this PR is getting to big.
