+++ lib/app_ipc.cpp
@@ -27,6 +27,7 @@
 #include "filesys.h"
 #include "miofile.h"
 #include "parse.h"
+#include "str_replace.h"
 #include "str_util.h"
 #include "url.h"
 #include "util.h"
