+++ lib/diagnostics_win.cpp
@@ -28,6 +28,7 @@
 #include "diagnostics.h"
 #include "error_numbers.h"
 #include "str_util.h"
+#include "str_replace.h"
 #include "util.h"
 #include "version.h"
