+++ lib/gui_rpc_client_ops.cpp
@@ -57,15 +57,16 @@
 #include <algorithm>
 #endif
 
-#include "common_defs.h"
 #include "diagnostics.h"
-#include "error_numbers.h"
-#include "md5_file.h"
-#include "miofile.h"
-#include "network.h"
 #include "parse.h"
 #include "str_util.h"
+#include "str_replace.h"
 #include "util.h"
+#include "error_numbers.h"
+#include "miofile.h"
+#include "md5_file.h"
+#include "network.h"
+#include "common_defs.h"
 
 #include "gui_rpc_client.h"
