+++ lib/hostinfo.cpp
@@ -32,8 +32,8 @@
 #include "error_numbers.h"
 #include "md5_file.h"
 #include "parse.h"
-#include "str_util.h"
 #include "util.h"
+#include "str_replace.h"
 
 #include "hostinfo.h"
