+++ lib/miofile.cpp
@@ -25,8 +25,9 @@
 #endif
 
 #include "error_numbers.h"
+#include "str_replace.h"
+
 #include "miofile.h"
-#include "str_util.h"
 
 using std::string;
