+++ lib/msg_log.cpp
@@ -23,6 +23,7 @@
 #endif
 
 #include "str_util.h"
+#include "str_replace.h"
 #include "util.h"
 
 #include "msg_log.h"
