+++ lib/notice.cpp
@@ -20,7 +20,7 @@
 #endif
 
 #include "error_numbers.h"
-#include "str_util.h"
+#include "str_replace.h"
 
 #include "notice.h"
