+++ lib/opencl_boinc.cpp
@@ -29,7 +29,7 @@
 #endif
 #include "miofile.h"
 #include "parse.h"
-#include "str_util.h"
+#include "str_replace.h"
 #include "util.h"
 
 #include "opencl_boinc.h"
