+++ lib/parse.cpp
@@ -42,6 +42,7 @@
 #include "boinc_stdio.h"
 
 #include "error_numbers.h"
+#include "str_replace.h"
 #include "str_util.h"
 #include "util.h"
