+++ lib/procinfo_unix.cpp
@@ -50,6 +50,7 @@
 #include "error_numbers.h"
 #include "filesys.h"
 #include "str_util.h"
+#include "str_replace.h"
 
 #include "procinfo.h"
