+++ lib/procinfo_win.cpp
@@ -1,9 +1,10 @@
 // much of this code is public-domain
 //
 #include "boinc_win.h"
-#include "diagnostics_win.h"
 #include "error_numbers.h"
+#include "diagnostics_win.h"
 #include "str_util.h"
+#include "str_replace.h"
 
 #include "procinfo.h"
