+++ lib/project_init.cpp
@@ -28,9 +28,9 @@
 
 #include "error_numbers.h"
 #include "filesys.h"
-#include "parse.h"
-#include "str_util.h"
 #include "url.h"
+#include "parse.h"
+#include "str_replace.h"
 
 #include "project_init.h"
