+++ lib/proxy_info.cpp
@@ -17,7 +17,7 @@
 
 #include "error_numbers.h"
 #include "parse.h"
-#include "str_util.h"
+#include "str_replace.h"
 
 #include "proxy_info.h"
