+++ lib/run_app_windows.cpp
@@ -20,14 +20,14 @@
 
 #include "boinc_win.h"
 
-#include "base64.h"
-#include "common_defs.h"
-#include "error_numbers.h"
-#include "filesys.h"
-#include "parse.h"
-#include "str_util.h"
-#include "util.h"
 #include "win_util.h"
+#include "filesys.h"
+#include "error_numbers.h"
+#include "str_replace.h"
+#include "common_defs.h"
+#include "util.h"
+#include "parse.h"
+#include "base64.h"
 
 HANDLE sandbox_account_interactive_token = NULL;
 HANDLE sandbox_account_service_token = NULL;
