+++ samples/wrappture/wrappture.cpp
@@ -41,12 +41,12 @@
 
 #include "boinc_api.h"
 #include "diagnostics.h"
-#include "error_numbers.h"
 #include "filesys.h"
 #include "parse.h"
-#include "proc_control.h"
 #include "str_util.h"
 #include "str_replace.h"
+#include "util.h"
+#include "error_numbers.h"
 
 #include "wrappture.h"
